# adcsProxy/src/main/java/de/trustable/ca3s/adcs/proxy/web/rest/AdcsCertificateBuilderApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.11).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package de.trustable.ca3s.adcs.proxy.web.rest;


import javax.validation.Valid;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import de.trustable.ca3s.adcs.proxy.web.dto.CertificateEnrollmentResponse;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2019-10-20T18:27:31.111Z[GMT]")
@Tag(name = "adcsCertificateBuilder", description = "the adcsCertificateBuilder API")
public interface AdcsCertificateBuilderApi {

    @Operation(summary = "request a certificate",
        description = "build a certificate using the local ADCS. The request parameter are wrapped into Javascript Web Signature",
        responses = {
        @ApiResponse(responseCode = "201", description = "new certificate created"),
        @ApiResponse(responseCode = "400", description = "invalid input, object invalid"),
        @ApiResponse(responseCode = "401", description = "authentication / authorization missing, especially JWS invalid") })
    @RequestMapping(value = "/adcsCertificateBuilder",
        produces = { "application/json" },
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<CertificateEnrollmentResponse> buildCertificate(@Parameter(description = "CSR and additional attributes (e.g. template) wrapped in a JWS"  )  @Valid @RequestBody JWSWrappedRequest body);

}